requirements.txt
simpletalsix>=6.2.0.dev3
requests>=2.13
werkzeug>=0.11
commonmark>=0.7.3